remove single responsbility from websocket integration
